refactor: move unit tests out to separate `ricer::test` module
BREAKING CHANGE: unit tests will not exist in the same area as production code
