+++ src/tests.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2024 Jason Pena <jasonpena@awkless.com>
+// SPDX-License-Identifier: MIT
+
+mod context;
+mod ui;
